+++ includes/Constants.php
@@ -38,6 +38,11 @@ final class Constants {
 	/**
 	 * @var string
 	 */
+	public const CONFIG_KEY_FACEBOOK_APP_ID = 'AchievementBadgesFacebookAppId';
+
+	/**
+	 * @var string
+	 */
 	public const PREF_KEY_ACHIEVEMENT_ENABLE = 'achievementbadges-beta-feature-achievement-enable';
 
 	/**