+++ modules/mw.achievementbadges.js
@@ -1,0 +1,4 @@
+(function () {
+  'use strict';
+  mw.achievementbadges = mw.achievementbadges || {};
+})();