+++ tests/phpunit/integration/Special/SpecialShareAchievementBadgeTest.php
@@ -2,13 +2,11 @@
 
 namespace MediaWiki\Extension\AchievementBadges\Tests\Integration\Special;
 
-// use MediaWiki\Extension\AchievementBadges\Achievement;
-// use MediaWiki\Extension\AchievementBadges\Constants;
+use MediaWiki\Extension\AchievementBadges\Achievement;
+use MediaWiki\Extension\AchievementBadges\Constants;
+use MediaWiki\Extension\AchievementBadges\Special\SpecialShareAchievementBadge;
 use SpecialPageTestBase;
-
-// use User;
-// use UserNotLoggedIn;
-// use Wikimedia\TestingAccessWrapper;
+use User;
 
 /**
  * @group AchievementBadges
@@ -18,17 +16,27 @@ use SpecialPageTestBase;
 class SpecialShareAchievementBadgeTest extends SpecialPageTestBase {
 
 	/**
-	 * Returns a new instance of the special page under test.
-	 *
-	 * @return SpecialPage
+	 * @inheritDoc
 	 */
 	protected function newSpecialPage() {
-		return new SpecialAchievements();
+		return new SpecialShareAchievementBadge();
 	}
 
-	public function testNothing() {
-		$user = $this->getTestUser();
-		$this->assertSame( 1, 1 );
+	public function testExecute() {
+		$key = 'share-badge-test';
+		$this->setMwGlobals( 'wg' . Constants::CONFIG_KEY_ACHIEVEMENTS, [
+			$key => [
+				'type' => 'instant'
+			]
+		] );
+		$user = new User();
+		$name = 'ShareAchievementTester';
+		$user->setName( $name );
+		$user->addToDatabase();
+		Achievement::achieve( [ 'user' => $user, 'key' => $key ] );
+
+		list( $html, ) = $this->executeSpecialPage( "$name/$key", null, 'qqx' );
+		$this->assertStringContainsString( 'special-shareachievementsbadge-message', $html );
 	}
 
 }